split interface and add new interface commun
