+++ apirestgenerator/apirest/src/main/java/com/Generator/apirest/core/design/ClassDesignInterface.java
@@ -1,8 +1,16 @@
 package com.Generator.apirest.core.design;
 
+import com.Generator.apirest.core.build.ParameterClassMethod;
 import com.Generator.apirest.services.builders.IImportModel;
 
-public interface ClassDesignInterface  extends IImportModel {
+import java.util.List;
+
+public interface ClassDesignInterface  extends DesingCommunInterface {
+
+
+
+
+
 
 
 }
