+++ apirestgenerator/apirest/src/main/java/com/Generator/apirest/core/design/MethodInterface.java
@@ -6,7 +6,7 @@ import com.Generator.apirest.services.builders.IBaseModel;
 
 import java.util.List;
 
-public interface MethodInterface extends IBaseModel {
+public interface MethodInterface extends DesingCommunInterface {
 
     // SyntaxOfMethod = annotation + modifiers + returnsType + methodName+ (parameter) + curlyBraces + methodBody + curlyBraces
 
@@ -22,38 +22,6 @@ public interface MethodInterface extends IBaseModel {
         return methodTx.toString();
     }
 
-
-    default public String annotationBuild(List<String> annotation){
-        StringBuilder methodTx = new StringBuilder();
-        if (annotation != null && annotation.size() > 0) {
-            annotation.stream().forEach(annotationOne ->{
-                methodTx.append(stringEnsamble(List.of(annotationOne,BREAK_LINE)));
-            });
-        } else {
-            methodTx.append(SPACE);
-        }
-        return methodTx.toString();
-    }
-
-
-    default public String parameterBuildStructure(List<ParameterClassMethod>parameters){
-        StringBuilder contexParameter = new StringBuilder();
-        if (parameters.size() > 0) {
-            contexParameter.append(PARENTHESES_OPEN);
-            for (int i = 0; i < parameters.size(); i++) {
-                String parameter = parameters.get(i).toString().replaceAll("\\s{1,2}"," ");
-                if (i < parameters.size() - 1) {
-                    contexParameter.append(stringEnsamble(List.of(parameter,COMMA)));
-                } else {
-                    contexParameter.append(parameter);
-                }
-            }
-            contexParameter.append(PARENTHESES_CLOSE );
-        } else {
-            contexParameter.append(PARENTHESES_OPEN_CLOSE);
-        }
-        return contexParameter.toString();
-    }
 
 
     default public String bodyBuildStructure(Boolean curlyBraces, String methodBody){
